+++ cvxpy/__init__.py
@@ -17,7 +17,7 @@ You should have received a copy of the GNU General Public License
 along with CVXPY.  If not, see <http://www.gnu.org/licenses/>.
 """
 
-__version__ = "0.2.24"
+__version__ = "0.2.28"
 from cvxpy.atoms import *
 from cvxpy.expressions.variables import Variable, Semidef, Symmetric, Bool, Int
 from cvxpy.expressions.constants import Parameter, CallbackParam, Constant
@@ -26,9 +26,10 @@ from cvxpy.problems.objective import Maximize, Minimize
 from cvxpy.problems.solvers.utilities import installed_solvers
 from cvxpy.error import SolverError
 from cvxpy.settings import (CVXOPT, GLPK, GLPK_MI,
-ECOS, ECOS_BB, SCS, GUROBI, ELEMENTAL,
+ECOS, ECOS_BB, SCS, GUROBI, ELEMENTAL, MOSEK,
 OPTIMAL, UNBOUNDED, INFEASIBLE, SOLVER_ERROR, ROBUST_KKTSOLVER,
 OPTIMAL_INACCURATE, UNBOUNDED_INACCURATE, INFEASIBLE_INACCURATE)
+from cvxpy.transforms import partial_optimize
 
 # Legacy names.
 from cvxpy.expressions.variables.semidef_var import Semidef as semidefinite