+++ cvxpy/expressions/constants/callback_param.py
@@ -33,3 +33,9 @@ class CallbackParam(Parameter):
         """Evaluate the callback to get the value.
         """
         return self._validate_value(self._callback())
+
+    def get_data(self):
+        """Returns info needed to reconstruct the expression besides the args.
+        """
+        return [self._callback, self._rows, self._cols,
+                self._name, self.sign_str]