+++ cvxpy/expressions/variables/symmetric.py
@@ -29,7 +29,7 @@ def Symmetric(n, name=None):
     """
     var = SymmetricUpperTri(n, name)
     fill_mat = Constant(upper_tri_to_full(n))
-    return types.reshape()(fill_mat*var, n, n)
+    return types.reshape()(fill_mat*var, int(n), int(n))
 
 def upper_tri_to_full(n):
     """Returns a coefficient matrix to create a symmetric matrix.
@@ -73,6 +73,11 @@ class SymmetricUpperTri(Variable):
     def __init__(self, n, name=None):
         self.n = n
         super(SymmetricUpperTri, self).__init__(n*(n+1)//2, 1, name)
+
+    def get_data(self):
+        """Returns info needed to reconstruct the expression besides the args.
+        """
+        return [self.n, self.name]
 
     def canonicalize(self):
         upper_tri = lu.create_var((self.size[0], 1), self.id)